Modified readByte method to return an integer (0-255) for the byte value and -1 if there is no byte at that position, instead of the earlier behaviour, which was to return a byte and throw an IndexOutOfBoundsException if the position was not valid.

+++ src/net/domesdaybook/reader/Reader.java
@@ -33,9 +33,10 @@ public interface Reader extends Closeable {
      * Read a byte from a given position.
      *
      * @param position The position of the byte to read.
-     * @return byte The byte at the position given.
+     * @return int The byte value at the position given as an integer (0-255)
+     *         If there is no byte at the position, it returns -1.
      */
-    byte readByte(final long position) throws IOException;
+    int readByte(final long position) throws IOException;
 
     
     /**
